[env](compiler) Reduce the compilation time of aggregate functions. (#61179)
Previously, some of our aggregate functions were implemented in very
large .cpp files, which made compilation slow.
For example, be/src/exprs/aggregate/aggregate_function_window.cpp takes
347.25 seconds to compile. T

his PR splits these files into multiple smaller files so they can be
compiled in parallel.

before
```
 347.25  be/src/exprs/aggregate/aggregate_function_window.cpp
```

now
```
[1/7] Compiling: be/src/exprs/aggregate/aggregate_function_window.cpp
  → 25.9s

[2/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_first.cpp
  → 60.5s

[3/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_funnel.cpp
  → 19.6s

[4/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_lag.cpp
  → 60.6s

[5/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_last.cpp
  → 63.9s

[6/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_lead.cpp
  → 69.1s

[7/7] Compiling: be/src/exprs/aggregate/aggregate_function_window_nth_value.cpp
  → 74.5s
```
