+++ be/src/exprs/aggregate/aggregate_function_min_max.cpp
@@ -20,150 +20,16 @@
 
 #include "exprs/aggregate/aggregate_function_min_max.h"
 
-#include "core/data_type/data_type.h"
-#include "core/data_type/data_type_nullable.h"
-#include "core/data_type/define_primitive_type.h"
-#include "core/types.h"
 #include "exprs/aggregate/aggregate_function_simple_factory.h"
-#include "exprs/aggregate/factory_helpers.h"
-#include "exprs/aggregate/helpers.h"
 
 namespace doris {
 #include "common/compile_check_begin.h"
-/// min, max
-template <template <typename> class Data>
-AggregateFunctionPtr create_aggregate_function_single_value(const String& name,
-                                                            const DataTypes& argument_types,
-                                                            const DataTypePtr& result_type,
-                                                            const bool result_is_nullable,
-                                                            const AggregateFunctionAttr& attr) {
-    assert_arity_range(name, argument_types, 1, 1);
-    switch (argument_types[0]->get_primitive_type()) {
-    case PrimitiveType::TYPE_STRING:
-    case PrimitiveType::TYPE_CHAR:
-    case PrimitiveType::TYPE_VARCHAR:
-    case PrimitiveType::TYPE_JSONB:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataString>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DATE:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_DATE>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DATETIME:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_DATETIME>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DATEV2:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_DATEV2>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DATETIMEV2:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_DATETIMEV2>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_TIMESTAMPTZ:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_TIMESTAMPTZ>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_TIME:
-    case PrimitiveType::TYPE_TIMEV2:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_TIMEV2>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_IPV4:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_IPV4>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_IPV6:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_IPV6>>>>(
-                argument_types, result_is_nullable, attr);
-    // For boolean, tinyint, smallint, int, bigint, largeint
-    case PrimitiveType::TYPE_BOOLEAN:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_BOOLEAN>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_TINYINT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_TINYINT>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_SMALLINT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_SMALLINT>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_INT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_INT>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_BIGINT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_BIGINT>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_LARGEINT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_LARGEINT>>>>(
-                argument_types, result_is_nullable, attr);
-    // For float, double
-    case PrimitiveType::TYPE_FLOAT:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_FLOAT>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DOUBLE:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataFixed<TYPE_DOUBLE>>>>(
-                argument_types, result_is_nullable, attr);
-    // For decimal
-    case PrimitiveType::TYPE_DECIMAL32:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataDecimal<TYPE_DECIMAL32>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DECIMAL64:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataDecimal<TYPE_DECIMAL64>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DECIMALV2:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataDecimal<TYPE_DECIMALV2>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DECIMAL128I:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataDecimal<TYPE_DECIMAL128I>>>>(
-                argument_types, result_is_nullable, attr);
-    case PrimitiveType::TYPE_DECIMAL256:
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<Data<SingleValueDataDecimal<TYPE_DECIMAL256>>>>(
-                argument_types, result_is_nullable, attr);
-    default:
-        return nullptr;
-    }
-}
 
-// any_value
-template <template <typename> class Data>
-AggregateFunctionPtr create_aggregate_function_single_value_any_value_function(
-        const String& name, const DataTypes& argument_types, const DataTypePtr& result_type,
-        const bool result_is_nullable, const AggregateFunctionAttr& attr) {
-    AggregateFunctionPtr res = create_aggregate_function_single_value<Data>(
-            name, argument_types, result_type, result_is_nullable, attr);
-    if (res) {
-        return res;
-    }
-    const DataTypePtr& argument_type = remove_nullable(argument_types[0]);
-    if (argument_type->get_primitive_type() == PrimitiveType::TYPE_ARRAY ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_MAP ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_STRUCT ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_AGG_STATE ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_BITMAP ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_HLL ||
-        argument_type->get_primitive_type() == PrimitiveType::TYPE_QUANTILE_STATE) {
-        return creator_without_type::create_unary_arguments<
-                AggregateFunctionsSingleValue<SingleValueDataComplexType>>(
-                argument_types, result_is_nullable, attr);
-    }
-
-    return nullptr;
-}
+// Template definition lives in aggregate_function_min_max_impl.h
+// Explicit instantiations provided by:
+//   aggregate_function_min_max_max.cpp  (AggregateFunctionMaxData)
+//   aggregate_function_min_max_min.cpp  (AggregateFunctionMinData)
+//   aggregate_function_min_max_any.cpp  (AggregateFunctionAnyData)
 
 void register_aggregate_function_minmax(AggregateFunctionSimpleFactory& factory) {
     factory.register_function_both(
@@ -171,8 +37,7 @@ void register_aggregate_function_minmax(AggregateFunctionSimpleFactory& factory)
     factory.register_function_both(
             "min", create_aggregate_function_single_value<AggregateFunctionMinData>);
     factory.register_function_both(
-            "any",
-            create_aggregate_function_single_value_any_value_function<AggregateFunctionAnyData>);
+            "any", create_aggregate_function_single_value<AggregateFunctionAnyData>);
     factory.register_alias("any", "any_value");
 }
