+++ be/src/exprs/aggregate/aggregate_function_min_max_min.cpp
@@ -1,0 +1,27 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+#include "exprs/aggregate/aggregate_function_min_max_impl.h"
+
+namespace doris {
+
+// Explicit instantiation for AggregateFunctionMinData
+template AggregateFunctionPtr create_aggregate_function_single_value<AggregateFunctionMinData>(
+        const String&, const DataTypes&, const DataTypePtr&, const bool,
+        const AggregateFunctionAttr&);
+
+} // namespace doris
