+++ be/src/exprs/aggregate/aggregate_function_topn.cpp
@@ -22,6 +22,7 @@
 
 #include "core/data_type/data_type.h"
 #include "core/data_type/define_primitive_type.h"
+#include "exprs/aggregate/aggregate_function_simple_factory.h"
 #include "exprs/aggregate/helpers.h"
 
 namespace doris {
@@ -42,51 +43,18 @@ AggregateFunctionPtr create_aggregate_function_topn(const std::string& name,
     return nullptr;
 }
 
-template <PrimitiveType T>
-using ImplArray = AggregateFunctionTopNImplArray<T, false>;
-template <PrimitiveType T>
-using ImplArrayWithDefault = AggregateFunctionTopNImplArray<T, true>;
-
-using creator =
-        creator_with_type_list<TYPE_TINYINT, TYPE_SMALLINT, TYPE_INT, TYPE_BIGINT, TYPE_LARGEINT,
-                               TYPE_FLOAT, TYPE_DOUBLE, TYPE_DECIMAL32, TYPE_DECIMAL64,
-                               TYPE_DECIMAL128I, TYPE_DECIMAL256, TYPE_VARCHAR, TYPE_DATEV2,
-                               TYPE_DATETIMEV2, TYPE_TIMESTAMPTZ, TYPE_IPV4, TYPE_IPV6>;
-
+// Forward declarations — implementations live in separate TUs
 AggregateFunctionPtr create_aggregate_function_topn_array(const std::string& name,
                                                           const DataTypes& argument_types,
                                                           const DataTypePtr& result_type,
                                                           const bool result_is_nullable,
-                                                          const AggregateFunctionAttr& attr) {
-    bool has_default_param = (argument_types.size() == 3);
-    if (has_default_param) {
-        return creator::create<AggregateFunctionTopNArray, ImplArrayWithDefault>(
-                argument_types, result_is_nullable, attr);
-    } else {
-        return creator::create<AggregateFunctionTopNArray, ImplArray>(argument_types,
-                                                                      result_is_nullable, attr);
-    }
-}
-
-template <PrimitiveType T>
-using ImplWeight = AggregateFunctionTopNImplWeight<T, false>;
-template <PrimitiveType T>
-using ImplWeightWithDefault = AggregateFunctionTopNImplWeight<T, true>;
+                                                          const AggregateFunctionAttr& attr);
 
 AggregateFunctionPtr create_aggregate_function_topn_weighted(const std::string& name,
                                                              const DataTypes& argument_types,
                                                              const DataTypePtr& result_type,
                                                              const bool result_is_nullable,
-                                                             const AggregateFunctionAttr& attr) {
-    bool has_default_param = (argument_types.size() == 4);
-    if (has_default_param) {
-        return creator::create<AggregateFunctionTopNArray, ImplWeightWithDefault>(
-                argument_types, result_is_nullable, attr);
-    } else {
-        return creator::create<AggregateFunctionTopNArray, ImplWeight>(argument_types,
-                                                                       result_is_nullable, attr);
-    }
-}
+                                                             const AggregateFunctionAttr& attr);
 
 void register_aggregate_function_topn(AggregateFunctionSimpleFactory& factory) {
     factory.register_function_both("topn", create_aggregate_function_topn);
